NewConversationActivity.java : Added the AutoContactFillAdapter instance for auto contact fill capability.
single_contact.xml :
Added TextField to render the contacts

AutoContactFillAdapter.java:
This is the adapter that is setup to work with the "newMsgRecipientAutoCompleteTextView"
Maybe changed later to adapt to any textfield we'd use for auto contacts fill.

ContactsUtil.java:
Added this class to provide static contact info from the application device.